update CosaUtilGetIfAddr() result data type ( ULONG -> uint32_t )
An IPv4 address is a 32bit value.

+++ custom/comcast/source/TR-181/custom_sbapi/cosa_deviceinfo_apis_custom.c
@@ -253,7 +253,7 @@ CosaDmlDiGetRouterIPAddress
 {
     UNREFERENCED_PARAMETER(hContext);
     char *interface = "erouter0";
-    uint32_t ip = (uint32_t) CosaUtilGetIfAddr (interface);
+    uint32_t ip = CosaUtilGetIfAddr (interface);
     unsigned char *a = (unsigned char *) &ip;
 
     /*
